FIX Solving comparison of int and size_t

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,4 +1,3 @@
-#include <stddef.h>
 #include "function_pointers.h"
 
 /**
@@ -13,7 +12,7 @@
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	int i;
+	size_t i;
 
 	if (!array || !action)
 		return;
